fix: refactor sync restart & fix wait initial sync race
