+++ helper/util/conn.go
@@ -1,11 +1,10 @@
 package util
 
 import (
+	"context"
+	"google.golang.org/grpc"
 	"io"
 	"net"
-	"time"
-
-	"google.golang.org/grpc"
 )
 
 // NewClientConnection creates a new client connection for the given reader and writer
@@ -13,7 +12,7 @@ func NewClientConnection(reader io.Reader, writer io.Writer) (*grpc.ClientConn, 
 	pipe := NewStdStreamJoint(reader, writer, false)
 
 	// Set up a connection to the server.
-	return grpc.Dial("", grpc.WithInsecure(), grpc.WithDialer(func(addr string, timeout time.Duration) (net.Conn, error) {
+	return grpc.Dial("", grpc.WithInsecure(), grpc.WithContextDialer(func(ctx context.Context, addr string) (net.Conn, error) {
 		return pipe, nil
 	}))
 }
